Rework the ViewModel of profiles and users
