+++ app/src/main/java/com/github/sdp/mediato/MainActivity.java
@@ -3,7 +3,6 @@ package com.github.sdp.mediato;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
@@ -16,8 +15,10 @@ import com.github.sdp.mediato.ui.ExploreFragment;
 import com.github.sdp.mediato.model.Review;
 import com.github.sdp.mediato.ui.HomeFragment;
 import com.github.sdp.mediato.ui.MyProfileFragment;
+import com.github.sdp.mediato.ui.ReadOnlyProfileFragment;
 import com.github.sdp.mediato.ui.SearchFragment;
-import com.github.sdp.mediato.ui.viewmodel.ProfileViewModel;
+import com.github.sdp.mediato.ui.viewmodel.MyProfileViewModel;
+import com.github.sdp.mediato.ui.viewmodel.ReadOnlyProfileViewModel;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.gson.Gson;
 
@@ -29,11 +30,12 @@ public class MainActivity extends AppCompatActivity implements FragmentSwitcher 
 
   ActivityMainBinding binding;
   MyProfileFragment myProfileFragment;
+  ReadOnlyProfileFragment readOnlyProfileFragment;
   SearchFragment searchFragment;
   ExploreFragment exploreFragment;
 
-  private ProfileViewModel currentUserViewModel;
-  private ProfileViewModel otherUsersViewModel;
+  private MyProfileViewModel myProfileViewModel;
+  private ReadOnlyProfileViewModel readOnlyProfileViewModel;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -41,12 +43,17 @@ public class MainActivity extends AppCompatActivity implements FragmentSwitcher 
     binding = ActivityMainBinding.inflate(getLayoutInflater());
     setContentView(binding.getRoot());
 
-    // Choose the default fragment that opens on creation of the MainActivity
-    setDefaultFragment();
-
     // Initialize the ViewModels
-    currentUserViewModel = new ViewModelProvider(this).get(ProfileViewModel.class);
-    otherUsersViewModel = new ViewModelProvider(this).get(ProfileViewModel.class);
+    myProfileViewModel = new ViewModelProvider(this).get(MyProfileViewModel.class);
+    readOnlyProfileViewModel = new ViewModelProvider(this).get(ReadOnlyProfileViewModel.class);
+
+    // Get the username set by the profile creation activity
+    String username = getIntent().getStringExtra("username");
+
+    myProfileViewModel.setUsername(username);
+
+    // Choose the default fragment that opens on creation of the MainActivity
+    setDefaultFragment(username);
 
     // Set the bottomNavigationView
     binding.bottomNavigationView.setBackground(null);
@@ -82,13 +89,12 @@ public class MainActivity extends AppCompatActivity implements FragmentSwitcher 
    * profile creation. Later this can be changed to show for example the Home Screen if the user is
    * already logged in.
    */
-  private void setDefaultFragment() {
+  private void setDefaultFragment(String username) {
     myProfileFragment = new MyProfileFragment();
+    readOnlyProfileFragment = new ReadOnlyProfileFragment();
     searchFragment = new SearchFragment();
     exploreFragment = new ExploreFragment();
 
-    // Get the username set by the profile creation activity
-    String username = getIntent().getStringExtra("username");
     Review review = new Gson().fromJson(
             getIntent().getStringExtra("review"), Review.class);
     Bundle args = new Bundle();
@@ -108,12 +114,12 @@ public class MainActivity extends AppCompatActivity implements FragmentSwitcher 
     replaceFragment(myProfileFragment);
   }
 
-  public ProfileViewModel getCurrentUserViewModel(){
-    return currentUserViewModel;
+  public ReadOnlyProfileViewModel getReadOnlyProfileViewModel(){
+    return readOnlyProfileViewModel;
   }
 
-  public ProfileViewModel getOtherUsersViewModel(){
-    return otherUsersViewModel;
+  public ReadOnlyProfileViewModel getMyProfileViewModel(){
+    return myProfileViewModel;
   }
 
   @Override
